Add files via upload
This code can be used in extension with the "Computing a 2D affine Toric variety from generators of the cone" to resolve singularities of an affine toric variety. 
The following assumptions have been made:
1. $e_2$ is a generator of the cone.
2. $m*e_1 - k*e_2$ is also a generator of the cone where $0 \leq k < m$. 
Note that any affine toric variety can be converted with an automorphism of the lattice such that the generators are of the above form.